scripts/requirements.txt
jsonschema
pyyaml
torch
twython
git+https://github.com/adapter-hub/adapter-transformers.git@v2
